Edryn Initial Sex Offer Requirements Change
If the player has a dick, but none are horse-type, then they can't proposition Edryn for sex for the first time unless they have a corruption of 50 or turned on low standards. This is a bit silly. I've changed the requirements to instead require that the player has at least 33 lust (the same requirement as most sex scene options)